+++ crates/matrix-sdk-ui/src/timeline/controller/mod.rs
@@ -274,7 +274,7 @@ pub fn default_event_filter(event: &AnySyncTimelineEvent, rules: &RoomVersionRul
                             UnstablePollStartEventContent::New(_),
                         )
                         | AnyMessageLikeEventContent::CallInvite(_)
-                        | AnyMessageLikeEventContent::CallNotify(_)
+                        | AnyMessageLikeEventContent::RtcNotification(_)
                         | AnyMessageLikeEventContent::RoomEncrypted(_) => true,
 
                         _ => false,