+++ crates/matrix-sdk-ui/src/timeline/event_item/mod.rs
@@ -616,7 +616,7 @@ impl EventTimelineItem {
             | TimelineItemContent::FailedToParseMessageLike { .. }
             | TimelineItemContent::FailedToParseState { .. }
             | TimelineItemContent::CallInvite
-            | TimelineItemContent::CallNotify => None,
+            | TimelineItemContent::RtcNotification => None,
         };
 
         if let Some(body) = body {